requirements.txt
streamlit
langchain
huggingface_hub
PyPDF2
faiss-cpu
sentence-transformers
diffusers  # Optional for vision
torch
torchvision
python-dotenv